corrijo metodo envio header
